Fix an Background-Image Problem
