+++ my-own-mentor-frontend/src/App.js
@@ -7,8 +7,19 @@ import {Switch, Route, BrowserRouter} from "react-router-dom";
 import {UserDispatchContext} from "./user/UserContext";
 import LandingPage from "./pages/LandingPage";
 import MentorAppBar from "./components/MentorAppBar";
+import makeStyles from "@material-ui/core/styles/makeStyles";
 
+const useStyles = makeStyles( () => ({
 
+    image:{
+        backgroundImage: 'url(https://cdn.pixabay.com/photo/2018/05/19/00/53/online-3412473_1280.jpg)',
+        backgroundRepeat: 'no-repeat',
+        backgroundSize: 'cover',
+        backGroundPosition: "center",
+        minHeight: '72vh',
+
+    }
+}));
 
 function Navigation() {
   const dispatch = useContext(UserDispatchContext);
@@ -43,9 +54,13 @@ function Navigation() {
 
 function App() {
 
+    const classes = useStyles();
+
   return (
   <UserContextProvider>
+      <div className={classes.image}>
    <Navigation />
+      </div>div>
   </UserContextProvider>
   );
 }
